Zmiana interfejsu LuvPixela, zmiana dabli na flołty

+++ src/main/java/mean/shift/kernel/Kernel.java
@@ -2,8 +2,8 @@ package mean.shift.kernel;
 
 public interface Kernel {
 
-	public double calculate(double distance, int range);
+	public float calculate(float distance, int range);
 
-	public double gFunction(double distance, int range);
+	public float gFunction(float distance, int range);
 
 }
