+++ src/main/java/mean/shift/kernel/RectangularKernel.java
@@ -3,13 +3,13 @@ import java.lang.Math;
 public class RectangularKernel implements Kernel {
 
 	@Override
-	public double calculate(double distance, int range) {
-		double x = Math.pow(distance/range, 2);
+	public float calculate(float distance, int range) {
+		float x = (float)Math.pow(distance/range, 2);
 		return (x < 0.5 && x > -0.5) ? 1 : 0;
 	}
 
 	@Override
-	public double gFunction(double distance, int range) {
+	public float gFunction(float distance, int range) {
 		return calculate(distance,range);
 	}
 }
