+++ src/main/java/mean/shift/processing/EuclideanMetrics.java
@@ -7,14 +7,14 @@ public class EuclideanMetrics implements Metrics {
 	}
 
 	@Override
-	public double getDistance(double ... points) {
+	public float getDistance(float ... points) {
 		if (points.length == 1)
 			return 0;
 		double sum = 0;
 		for (int i = 0; i < points.length; i++) {
 			sum += points[i] * points[i];
 		}
-		return Math.sqrt(sum);
+		return (float)Math.sqrt(sum);
 	}
 
 }
